test: fix two CI failures — timing thresholds and float promotion
Both were pre-existing and both needed a real reproduction rather than a
guess. To get one I installed llvm@18 locally, matching the clang-18 the
Ubuntu job uses, and built the whole tree with it.

## test_when_all.cpp — wall-clock thresholds, not framework bugs

`when_all: real parallelism` asserted `elapsed < 130` against a ~150ms
sequential cost. Under ASan/UBSan a loaded runner took 169ms doing
correctly parallel work. Nothing about "runs in parallel" implies
"finishes within 130ms on every machine", so the threshold was the
defect.

Now the test measures a sequential baseline on the spot and requires the
parallel run to come in under 80% of it. The budget scales with whatever
the host and its sanitizers cost, and the assertion still means
something: forcing sequential awaits reports 159ms against a 140ms
budget, i.e. it fails as intended. `INFO` logs all three numbers so a
future CI failure explains itself instead of just showing two integers.

`when_any: fastest among real-time waiters wins` used 100/20/60ms and
picked the 100ms task as winner on CI (`CHECK(1 == 99)`). A 40ms gap is
not an ordering signal once instrumentation and scheduler contention
push wakeup jitter into the same range. Widened to 20/300/500ms: the
winner is now unambiguous unless the machine stalls for a quarter
second, which would be real news. Runtime is bounded by the winner
(~20ms) plus the drain wait, so the suite barely notices. Added
`REQUIRE(done.load())` so "wrong winner" and "timed out" are
distinguishable, and raised the deadline accordingly.

Verified: ASan 12/12; then 24 spinning processes on 12 cores (to mimic a
contended runner) and the full async suite five times over — 135/135
every run. Under the old thresholds this environment is exactly what
broke them.

## -Wdouble-promotion — the promotion is at the call site

`float == doctest::Approx` promotes the left operand, because `Approx`
holds a double. clang-18 on Ubuntu reports it from inside doctest's
comparison template, where no local edit can reach. Dropping the `f`
suffix does nothing: the float-typed *left* operand is what gets
promoted, so the fix is to widen explicitly at the call site.

Three sites in two files. AppleClang does not diagnose this at all,
which is why it needed the llvm@18 build to check — noting for the
record that a local libc++ build still does not reproduce the Ubuntu
libstdc++ diagnostic, so CI remains the final word on this one.

Verified: debug 12/12 zero warnings, ASan 12/12, and a full clang-18
build of the tree at 8/8 with zero warnings.

+++ modules/binding/include/aria/binding/testing/adapter_conformance.hpp
@@ -261,7 +261,11 @@ inline void run_float_two_way_via_double_view(Harness& h) {
     auto adapter = h.make_adapter();
 
     adapter->set_float(v, 2.5f);
-    CHECK(adapter->get_float(v) == doctest::Approx(2.5f));
+    // Widen explicitly: `Approx` holds a double, so comparing a float
+    // against it promotes implicitly, and -Wdouble-promotion flags that
+    // from inside doctest's comparison template. See the same note in
+    // test_binding_readonly_source.cpp.
+    CHECK(static_cast<double>(adapter->get_float(v)) == doctest::Approx(2.5));
 }
 
 // ═══════════════════════════════════════════════════════════════════════
